Add comment to component
Comment is added outside of the @Component decorator so that
the comment can be used
in the docs property of the docs-json or docs-vscode output targets.
This will also prevent the system from populating the docs property
using content between the main title and first subtitle in the
readme.md file which might not
accurately describing the component.