fix(jest): move svgTransform to .cjs to support require() in ESM context
Since the component library uses `"type": "module"` in package.json, Jest treats all
`.js` and `.ts`	 files as ES modules. This caused `svgTransform.js` to throw a
Reference Error that require was not defined in ES module scope due to the use of
CommonJS `require()`.

To resolve this, svgTransform was renamed to `svgTransform.cjs`, and the Jest config
was updated to point to this new `.cjs` file, allowing it to be interpreted correctly
as CommonJS without affecting the rest of the project's ESM configuration.